Cr50: Enable normal sleep, too
This adds support for the "sleep" low-power mode. It consumes
less power than simply waiting, but doesn't require a full warm
boot to resume.

BUG=chrome-os-partner:49955
BRANCH=none
TEST=make buildall; test on Cr50

Configure the Cr50 to sleep when idle (refer to previous commit
messages for the setup required).

On the console, use the "idle" command to test the three
different modes:

  idle w      - wfi, wakes instantly
  idle s      - sleep, wakes slowly but without rebooting
  idle d      - deep sleep, wakes via warm boot

You can tell the difference between wfi and sleep by observing
that the first character is lost when typing on the serial
console while in sleep (remember that it will wait at least 10
seconds after the last console input before sleeping).

Signed-off-by: Bill Richardson <wfrichar@chromium.org>
Change-Id: Ib2584aa44ab885f0c8369ec938ee17b935aa0898