Cr50: SPI tests need to poke the target to wake it up
If the Cr50 is sleeping, it can't wake up quickly enough when
traffic arrives on the SPI bus to catch all of the transmitted
data. This makes the FTDI framework poke the Cr50 by asserting
SPS_CS_L briefly, then waiting a few milliseconds before starting
the tests.

BUG=chrome-os-partner:49955
BRANCH=none
TEST=make buildall; test on Cr50

Ensure that the Cr50 invokes sleep or deep sleep when idle (refer
to previous commit messages for the setup required), then

  cd test/tpm_test
  make
  ./tpmtest.py

Before this CL, the test hung or failed because it couldn't get a
quick response from the Cr50. With this CL, the Cr50 wakes up and
the test passes.

Change-Id: I581475726313981a780beaaa37638e9c3b9ebec5
Signed-off-by: Bill Richardson <wfrichar@chromium.org>