common: flash: Add flash_block_and_read().
This commit adds a new function, flash_block_and_read(), which will
perform a flash_read, but will not try to yield or delay between
transactions.

All other uses of flash_read() remain unchanged and they will yield if
task switching has been started.

This also helps to fix an issue with when
CONFIG_REPLACE_LOADER_WITH_BSS_SLOW is enabled.  If we disable
interrupts before re-loading the loader from flash, the existing
flash_read() will try to sleep in between transactions, causing a hard
fault.  The hard fault is avoided by not allowing the read from flash to
try and sleep.

BUG=chrome-os-partner:51951
BRANCH=None
TEST=make -j buildall tests
TEST=Build and flash sentry; from userspace, issue the command 'ectool
reboot_ec RO' and 'ectool reboot_ec RW' and verify that the AP does not
reset and no panic is shown on the EC console.

Change-Id: I888c6408e447950d276a7988176afac81e2e44b9
Signed-off-by: Aseda Aboagye <aaboagye@google.com>