Update patch set 1
Patch Set 1:

Currently, I don't believe so. I've been trying to think of a way we might do that, but nothing so far.

I'm also not sure about how I feel with littering all the SPI functions with this yield parameter...just to prevent calling sleeps for this one case.

Patch-set: 1