Update patch set 1
Patch Set 1:

One other thing that I was going to mention was that, with the patchset that I uploaded, I had missed some sleeps in the spi_* functions, but I didn't run into this issue. Meaning that, it's not just the sleep which was causing the issue.

Even stranger is that I can no longer reproduce the issue.

Patch-set: 1