Update patch set 2
Patch Set 2:

I think the semantics of this are off. We really want to say the following:

task_start_synchronous() (different name suggestions welcomed)

which is disable irqs
disable scheduling and events.

The reason is that you want something to completely run to completion. And in this case we absolutely have to do that.

Making it asymmetric (i.e. no disable primitive) means people can't use it because if they do it will break the EC entirely. Since this construct would be employed on in the jump path that's A-OK.

Patch-set: 2
Label: Code-Review=0