Update patch set 1
Patch Set 1:

This is the 1st CL for the A/B NvMem partition. As stated in the CL description, the main functions are not implemented here, but the basic framework is in place (new files, calls from TPM, some of the #defines). There are 2 more CLs that build on this, the next one adds read/write/commit functionality and the next one sets up and toggles between partitions.

Patch-set: 1