Update patch set 2
Patch Set 2:

> During we update EC firmware by flashrom on dut, this cause would
 > occurs, flashrom would update RW first.

The RO and RW need to be flashed externally first for such a feature change.I don't think papering over how it was lucky the original RO even worked is appropriate since the lesson is that one needs to ensure RO and RW match w/ the hardware. In the case you described you were using an RO fw which was inherently incompatible with the hardware itself.

Patch-set: 2