Update patch set 3
Patch Set 3:

> How is this different than cprintf("%.*h", len, buf)?  I suppose in
 > that it can dump a large amount of data (but beware uart output
 > buffer size limits).
 > 
 > I don't think this belongs in uart.h.  That's the interface to the
 > low-level uart driver.  It'd be less weird in util.h.

cprintf would do, but it's convenient to have formatting for visual inspection.

I took a look at moving the function to util.c, but then util.h will have to include uart.h.  Is that an acceptable dependency?  (or I could abandon this change, as I'm the only one in need of it).

Patch-set: 3