Cr50: Add console command "usb" for testing
This adds a command to connect or disconnect, and to switch the
PHY between A or B.

BUG=chrome-os-partner:52055
BRANCH=none
TEST=make buildall; test on Cr50

Using a test board with both PHYs plugged in, try the various
commands:

  usb off
  usb on
  usb a
  usb b

The on/off option connects and disconnects, the a/b option
switches between PHYs. You can see the state change on the
console, or by running dmesg on the host.

Change-Id: I4c77e9c586ce197dc99b0b50af7396c253a1a377
Signed-off-by: Bill Richardson <wfrichar@chromium.org>