Update patch set 1
Patch Set 1:

Shawn, doing it in a separate shell session seams counter productive to me, the hassle of determining the pty, of keeping a terminal session just for this, etc. Being able to hit 'arrow up/return' in the same window seems much more convenient, note that the guys in "gameboy" running their own ec firmware were this double toggle is added.

Randall, sure, I can do it as a separate command, even though this seems a great overkill to me. What is the function of 'powerbtn' anyways - isn't it to reboot the AP? 

If so - how about forcing this command to do the right thing using compile time flags, not requiring an extra command line parameter?

Patch-set: 1