Update patch set 1
Patch Set 1:

powerbtn is to simulate power button presses.

It can be used to power the system on, force-power the system off (pwrbtn 8000), log the user out (pwrbtn 4000), or anything else that the power button can do.  It's used by FAFT testing.

We use 'apreset' and 'apshutdown' to reset or shut down the AP directly.

Patch-set: 1