skylake: de-assert RTCRST when system power up
always de-aseert RTCRST when system attempts to power up.
this is to handle the case where RTCRST is asserted manually (via
ectools for example) and caused the system to shutdown, RTCRST
will stay asserted and block the system to power up again.

BUG=none
BRANCH=none
TEST=use ectool or EC console to assert RTCRST signal, wait for
system to shutdown, press power button and system is able to boot
back to OS.

Change-Id: I1754f7d2855df0dab4d83cb931b3c724e05bdc9e
Signed-off-by: Kevin K Wong <kevin.k.wong@intel.com>