g: Add USB_STREAM_CONFIG_FULL macro to usb-stream.h
It's handy to use the usb-stream interfaces to avoid a lot of
typing. But not all the endpoints are traditional serial ports.
This just adds a new macro that lets us specify additional
parameters.

BUG=chrome-os-partner:50707
BRANCH=none
TEST=make buildall; test on Cr50

Verified that all the previous endpoints still work as before.
There are no endpoints that use the new macro yet.

Change-Id: Ia37901cbe3adc4a4650ab91db3596efa15a110de
Signed-off-by: Bill Richardson <wfrichar@chromium.org>