cr50: [DO NOT COMMIT] flash counter prototype
The persistent counter is composed of a low and high page.  The low page
bits are continuously struck (4 bits at a time) until a max of 4096 has
been counted.  Once the low page reaches max, the high page increments
by 2, and the low page is erased and begins counting again.  There are
various mechanisms in place to guard against accidental or intentional
interruptions, including the flash brownout mechanism.

BUG=none
BRANCH=none
TEST=auto increment counter overnight run

Change-Id: Iddcf0285e2c8a12798f5515df02b2d793833144a
Signed-off-by: Timothy Chen <timothytim@google.com>