Update patch set 1
Patch Set 1:

We need to be a bit careful about this one because it will (a) violate the USB-C spec and (b) prevent video out over USB-C from working.

Do we have a way to mark it so it does not remain in the code once we have got past development? Maybe we just need a bug (I don't think 50700 quite covers it) to ensure it is pulled out eventually?

Patch-set: 1
Reviewer: Gerrit User 1001164 <1001164@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0