Driver: BD99955: Prepare charger to enter Pseudo G3
In Pseudo G3 if both VBUS_EN & VCC_EN are disabled, on plugging the AC
on either of the ports, ACOK pin is not asserted as the charger cannot
detect the voltage on either of the ports. Hence the device cannot boot.
Call this function at board level functions before entering pseudo G3.

BUG=none
BRANCH=none
TEST=Manually tested on Amenia. Enter PG3, plug in the AC, charger ACOK
     is asserted and the device boots to S0.

Change-Id: Idbae2306948a470592219edfe475dc9b45c1df58
Signed-off-by: Vijay Hiremath <vijay.p.hiremath@intel.com>