Driver: BD99955: Prepare charger to enter Pseudo G3
In Pseudo G3 if both VBUS_EN & VCC_EN are disabled, on plugging the AC
on either of the ports, ACOK pin is not asserted because the charger
cannot detect the voltage on either of the ports. Hence the device cannot
boot until the power button is pressed.
Call this function at board level functions before entering pseudo G3
so that the VBUS_EN & VCC_EN are enabled in pseudo G3 and the board can
boot to S0 on plugging the AC on either of the ports.

BUG=none
BRANCH=none
TEST=Manually tested on Amenia. Enter PG3, plug in the AC, charger ACOK
     is asserted and the device boots to S0.

Change-Id: Idbae2306948a470592219edfe475dc9b45c1df58
Signed-off-by: Vijay Hiremath <vijay.p.hiremath@intel.com>