make: allow to see boards failed when making 'buildall'
It is quite annoying to see an error reported by buildall, but not
knowing what board(s) actually failed to build.

Create file for each board being built, put build progress information
in it, and remove it if the board build succeeded. Then, once the
build is completed see the failed boards in .failedboards/, if any,
and the states they failed at.

BRANCH=none
BUG=none
TEST=verified that the directory is empty after a successful buildall
     run, and has files named after failed boards in case of build
     failures.

Change-Id: I67eb1671cadf58d9f8feccebfcc860524f33c2a0
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>