cr50: Enable RDD
Enable RD Detection on cr50. When a debug accessory is detected by cr50,
the USB controller switches to using the CCD PHY and switches back to
the AP PHY when the cable is disconnected. This change also enables
controlling the UART TX output for the AP and EC console.

BUG=chrome-os-partner:50700
BRANCH=none
TEST=Attach suzy Q and verify the PHY is set to B. Detach Suzy Q and
check the phy has switched to A.

Change-Id: I07f4c4fb7f765788a0914dfc2451a11af8bf0ab2
Signed-off-by: Mary Ruthven <mruthven@chromium.org>