WIP: g: i2cs driver
A fairly detailed description of the driver is included in the
comments in the file.

As of this submission the setup is fairly brittle, as clock stretching
is not yet enabled, and there is no guarantee that the slave will
prepare the data for the master in time.

More testing will be required and enhancements will be added in the
future.

BRANCH=None
BUG=chrome-os-partner:40397
TEST=with the rest of the patches applied, an i2c app running on the
     desktop allows read and write TPM registers using the
     USB-FTDI-I2C cable connected to the B1 board.

Change-Id: I117850753c58cf374e152027ab6caa30833c3360
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>