lid: serialize lid open and close commands
This patch serializes the S3 lid open and close behavior.

Before lid open is initiated, we check if we are in the desired state
for lid open to work i.e. S3, S0iX, S5 or G3. If we are there, we send
the event immediately; otherwise we do a deferred check until we reach
the desired state.

The same sequence is done for lid close as well.

After retrying a fixed number of retries, we give up on a particular
desired end state; In such cases, we will also not process any
pending events.

BUG=chrome-os-partner:51005
BRANCH=none
TEST=loop dut-control lid_open:no sleep 1.5 lid_open_yes should
     work for 100+ iterations. Various delays from 1.0 to 3.0
     should be tried. S3, S5 and S0ix flows should all work.

Change-Id: Ie332323e7bf4d3093f9d7dfb32d9e5122edf1eb6
Signed-off-by: Kyoung Kim <kyoung.il.kim@intel.com>
Signed-off-by: Shobhit Srivastava <shobhit.srivastava@intel.com>
Signed-off-by: Jenny TC <jenny.tc@intel.com>
Signed-off-by: Archana Patni <archana.patni@intel.com>
Signed-off-by: Subramony Sesha <subramony.sesha@intel.com>