lid: serialize lid open and close commands
This patch serializes the S3 lid open and close behavior.

When lid close is initiated, we wait for the desired end state
to be realized i.e. S3, S0ix or S5 If the next event, lid open,
comes while the previous event is being processed, then we mark
it as pending and send it after the desired end state has been
reached for lid close.

The same sequence is done for lid open as well.

After retrying a fixed number of retries, we give up on a particular
desired end state; this can happen if there is a kernel abort or something
else which can cause the state not to be reached. In such cases, we will
also not process any pending events.

BUG=none
BRANCH=none
TEST= loop dut-control lid_open:no sleep 1.5 lid_open_yes should
      work for 100+ iterations. Various delays from 1.0 to 3.0
      should be tried. S3, S5 and S0ix flows should all work.

Change-Id: Ie332323e7bf4d3093f9d7dfb32d9e5122edf1eb6
Signed-off-by: Kyoung Kim <kyoung.il.kim@intel.com>
Signed-off-by: Shobhit Srivastava <shobhit.srivastava@intel.com>
Signed-off-by: Jenny TC <jenny.tc@intel.com>
Signed-off-by: Archana Patni <archana.patni@intel.com>
Signed-off-by: Subramony Sesha <subramony.sesha@intel.com>