Update patch set 1
Abandoned

I'm abandoning this CL.

I think it's ok to have a bigger number here since we're just using this struct as a layout. DMA1 can have smaller or bigger number of channels than DMA2. For example stm32f373 has 7 channels on DMA1 and 5 on DMA2. So, either way, we'll not be able to define stm32_dma_regs accurately.

Patch-set: 1
Status: abandoned