Update patch set 3
Patch Set 3:

The flow was changed - earlier, we were sending a command and waiting for a desired end state before sending the next command. 

Now before the command is issued, the desired start state is checked. If the state is already there, then the command is issued immediately. Otherwise a repeated deferred check is done and command issued when the state is reached.

Patch-set: 3