Update patch set 2
Patch Set 2:

This issue is still looming... (see https://chromium-review.googlesource.com/#/c/340273/)

[copying comment that I made on the bug]
I had investigated a bit further and was trying another approach to disable scheduling and faking events(I think I called it task_disable_scheduling() with no symmetric enable), but I still encountered exceptions (although, in the loader and sysjumping very quickly back and forth). Basically, the current way we disable interrupts for the Cortex-M4 is by clearing the bit in the PRIMASK register. When the EC makes a supervisor call (SVC), since interrupts are disabled, it gets escalated to a hard fault since the the SVC is a configurable priority interrupt. The alternative that I was trying was using BASEPRI to disable interrupts. When BASEPRI is non-zero, it prevents all exceptions with the same or lower priority from running. So, I set this to 1 allowing only the SVC to run, faked events, and sprinkled some reloading of the watchdog in some of the SPI transaction related functions. It still doesn't prevent code from accessing code in that region, as we currently don't have any sort of memory protection.

I don't recall exactly what exceptions that I was running into, and I haven't had time to go back and debug. It may have been the watchdog expiring.

Patch-set: 2