Update patch set 1
Patch Set 1:

I understand the reason for preventing splitting strings, but I think it should be for the coder to decide how to proceed if a string gets too long. Be it defining it at a different indent or splitting it such that search is still possible (say on a format statement).

This checkpatch change was accepted in coreboot.

Patch-set: 1