system: Disable task switching when reloading LFW.
CONFIG_REPLACE_LOADER_WITH_BSS_SLOW was introduced as a way to regain
some space by replacing the little firmware loader, which is only used
to load the current image, with some extra variables that would be in
the .bss section.

However, in the event of a sysjump, the EC needs to reload the little
firmware loader back from flash into RAM, in order to perform the
jump. This tries to happen as late as possible in the sysjump
process. The problem is such that once the loader is loaded back into
RAM, occasionally some code may modify some variables that were in that
region, thereby corrupting the loader which then leads to an exception
when we try to jump to the new image. This is because we currently
disable interrupts AFTER loading the loader.

This commit disables task scheduling right before loading the LFW back
into memory in order to minimize the chance of other code writing to
that location.

BUG=chromium:606035
BUG=chrome-os-partner:51951
BUG=chrome-os-partner:51438
BRANCH=None
TEST=make -j buildall tests;
TEST=Enable CONFIG_REPLACE_LOADER_WITH_BSS_SLOW on Sentry.  Using a
script, sysjump RO, wait 2 seconds, sysjump RW, wait 2 seconds, and
repeat for 500 iterations.

Change-Id: I2c1fe80b1ebbf900d4a5e82da809214f18d10248
Signed-off-by: Aseda Aboagye <aaboagye@google.com>