Update patch set 3
Patch Set 3: Verified+1

Okay, this is the approach that I was talking about. Basically, this adds an API (with no re-enable so people can't abuse this) which aims to disable task scheduling.

Basically, set a flag which will prevent the SVC handler from actually doing anything including calling out to __switch_to.

Additionally, use BASEPRI to disable interrupts instead of PRIMASK, so that making a supervisor call doesn't get escalated to a hard fault.

It also keeps the damage local to task.c without having to reach out and muck with a bunch of other files.

Patch-set: 3
Label: Verified=+1