skylake: add wake mask programming in EC
The suspend/shutdown sequence is the following:
(Kernel) (Coreboot SMI) (EC)

Earlier the wake masks were programmed in Coreboot. If the lid open event
is initiated after Coreboot programs the wake masks, then the PCH is pulsed
but system is still in S0. EC will complete the transition to S3 after this
and the system will not wake up.

As part of a patchset, the wake mask programming is being removed from Coreboot
and added to EC.

BUG=chrome-os-partner:51005
BRANCH=None
TEST=loop dut-control lid_open:no sleep 1.5 lid_open_yes should
     work for 100+ iterations. Various delays from 1.0 to 3.0
     should be tried. S3 and S5 flows should all work.

Change-Id: I2802c773daba3f425cc9dc539f3841e15990d790
Signed-off-by: Archana Patni <archana.patni@intel.com>
Signed-off-by: Subramony Sesha <subramony.sesha@intel.com>
Signed-off-by: Jenny TC <jenny.tc@intel.com>