Update patch set 1
Patch Set 1:

> sure thing, thank you for the suggestion.
 > 
 > Feel free to open a bug with your recommendations and assign it to
 > me.

Will do, I didn't initially because this alternate code path in uart_buffering was complicating a refactor.  I initially started to make the above suggested fix to polling_uart.c, but when I realized it didn't work I opted for removing it.  If you want to fix it that would be great.  If you don't think you'll be able to get to it today I may submit a CL for the proposed fix, but not the de-bitrotting, so that my refactor is unblocked.

Patch-set: 1