Update patch set 1
Patch Set 1: Code-Review+1

A side-benifit of this change is that we can remove adc_read_all_channels(). We don't actually read all ADC channels at once on any TOT board (though we did on snowball for a while, before the design changed). There is some efficiency loss in reading ADC channels individually vs all at once on stm32 (but not so on other chips), but since we aren't actually using it, I think it's ok to remove.

Patch-set: 1
Label: Code-Review=+1