WIP: Console: Replace one off queue implementation with common queue code
This is not done, but the basic idea is to have two queues, the primary
queue replaces the hand coded buffer management, and the secondary queue
is driven by a queue policy used by the primary queue to always point to
the portion of the primary buffer that is still valid to read, even if
the primary queue has already advanced its head (but not yet clobbered
the elements by advancing its tail).

There are some race conditions to deal with, but no more so than the
existing code, and it greatly simplifies the host command implementation.

Signed-off-by: Anton Staaf <robotboy@chromium.org>

BRANCH=None
BUG=None
TEST=make buildall -j

Change-Id: I91239023b5218d93ccf85ffe974339c412df8789