Update patch set 2
Patch Set 2:

So, the goal is to replace all of the input-related host events with a more generic input event?  Does this also include POWER_BUTTON, KEY_PRESSED, KEYBOARD_RECOVERY, KEYBOARD_FASTBOOT?

This seems like a major refactor with impact on FAFT testing.  We'll also need to update depthcharge to understand the new input events.  

Is there a design doc for this?

Patch-set: 2