Cr50: Enable USB suspend on the HID interface
The HID interface descriptor needs to specify its subclass and
protocol as zero in order for USB suspend to be triggered
properly.

In addition, crbug.com/431886 seems to still be present, so this
adds a workaround for that until we know more. We may need to
reopen that bug.

BUG=chrome-os-partner:52810
BRANCH=none
TEST=make buildall; try it on Cr50

Switch a Chromebook to dev-mode, disable rootfs verification.

Add USB device "18d1:5014" to the whitelist in
  /etc/laptop-mode/conf.d/board-specific/usb-autosuspend-cros.conf

Connect the USB analyzer between the Cr50 test board and the
Chromebook, observe that USB suspend occurs with this CL.

Change-Id: Ie86d95060934613511caa6e49d87b42613664427
Signed-off-by: Bill Richardson <wfrichar@chromium.org>