Update patch set 3
Patch Set 3:

> There was an EC reset after 22 iterations. I did not have servo
 > connected but the exception seems to be similar as before. In the
 > kernel console, I see flashrom starting the 23rd iteration but the
 > console becomes unresponsive before "Erasing and writing flash
 > chip" is printed. The system boots up fine after the exception.

Looks like I need to do some more analysis of the failure. It would be nice to know exactly at what point the exception occurred: prior to jumping to the loader, within the loader, or after starting to execute in the new image.

Although, if it froze before printing "Erasing and writing flash chip", that means it wasn't in a sysjump, correct? I'll need to double check on that one.

Patch-set: 3