Update patch set 3
Patch Set 3:

In my testing before but not with this patch, I believe I have seen all of the three failures that you described. The commonality among these failures is that the EC image in flash does not get corrupted and system reboots successfully. I believe I hit one of these cases in my test yesterday but not sure which one.

To determine, I will reproduce the issue on my kunimitsu with a servo connected and provide the logs.

Patch-set: 3