g: enable fast brownout detect while flash is being written
The flash module is more sensitive to voltage drops than the rest of
the chip, and if not properly processed, flash can be even damaged.

This patch enables the fast brownout detection ability for the
duration of flash write operations. As of now this would cause the
chip reset, it is not expected to happen, more of a precaution than
anything else.

BRANCH=none
BUG=chrome-os-partner:52581
TEST=with this code in place ran firmware upgrade operation a few
      times, no issues have been observed.

Change-Id: I435bbce73a12143a54dd5d76cb7c954edff52b4e
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>