cr50: monitor state of servo, ec and ap
There are a couple of issues that cr50 has when it cannot know the state
of servo, the EC, and the AP. This change adds support so we can detect
when the AP or EC has been powered on and when servo has been connected.
It uses the UART RX signals to know the power state of the AP and EC.
The TX signals are used to know when servo is attached.

BUG=chrome-os-partner:52056,chrome-os-partner:52322
BRANCH=none
TEST=Verify ec, ap, and servo interrupts are fired when they are powered
on or attached.

Change-Id: Ieb4218e630765b5f487beb2ac8cdbaf6ac2a6baf
Signed-off-by: Mary Ruthven <mruthven@chromium.org>