Update patch set 1
Patch Set 1:

We currently do pet the watchdog at the beginning of LFW. And with this patch series, we pet the watchdog prior to jumping to the loader, and now fairly close to the end of the LFW. We also pet it as part of our inits in the new image.

But like I mentioned before, we still watchdog in the beginning of the new image (even after we've pet the watchdog again).

Patch-set: 1