Update patch set 2
Patch Set 2: Code-Review-2

(7 comments)

Not quite there. I think we need to do autoconnect/disconnect a bit better. For the EC/AP, a rising edge instantly indicates "powered", but a falling edge must be debounced before we can tell if it's unpowered. We can kind of assume that the AP and EC will be unpowered at boot, if only because rebooting Cr50 will reboot them, but it would be better to use the same logic to reliably determine their state at boot.

We also need to handle the case where only the EC is powered. If/when the AP later powers up too, we want it to mirror the "enabled" state of the EC. 

Servo may already be connected at boot, so we'd never see a rising edge. It may be simpler to make the interrupts level sensitive instead. And if we've already enabled UARTs and *then* we connect servo, we want to disable them immediately, and be prepared to notice if servo is later disconnected.

Patch-set: 2
Label: Code-Review=-2