Update patch set 4
Patch Set 4:

Hi Shawn:
Sorry for late update. We try very hard to duplicate the symptom in our environment and finally did it.

You are right, there is no need for 'volatile'.
The reason for the issue seen is not related to the timer reading but to wake-up from deep sleep. Upon waking-up from deep sleep, the core execution may be disrupted. A workaround to this issue is added in clock.c. Please see commit message for detail.

Patch-set: 4
Label: Code-Review=0