EC Unit Test: Improvements to mutex test
Fixed an issue with the simple contention test. The test intends
to have task MTX2 grab the lock then have task MTX1 attempt to
obtain the lock. However, task MTX2 would grab then release the
lock prior task MTX1 trying to get the lock.

In addition, a new test was added to specfically test locking
and release order. The test uses the existing random_task()
function and verifies that all 3 tasks are pending on the lock being
released. Following the lock release, the test also verifies that
the pending task are run in highest to lowest priority order.

The issue description notes issues with the mass lock-unlocking
test related to random time delays being used when run on the
emmulator. Given that these tests are occasionally run on real
target hardware, this test was not modified so that existing
functionality was not modified.

BRANCH=none
BUG=chromium:611086
TEST=Manual
Ran 'make runtests' then ./build/host/mutex/mutex.exe to view
console output. Verified that the tests pass and console log
matched the expected output.

Change-Id: I6be7fc3b7176728f85953ca840df3f48da8c749a
Signed-off-by: Scott <scollyer@chromium.org>