Update patch set 1
Patch Set 1:

Unused data memory is used as a shared buffer. If the size of the shared buffer is too small, the hashing routine cannot claim a buffer large enough, and hash calculation will be delayed or fail. I believe that is the root cause of your problem, and if so, it is related to data RAM space.

Patch-set: 1