EC: Added mutex that allows multiple lock attempts from same task
Added 'sloppy_mutex' which is nearly identical to existing mutex
lock/unlock functions, but allows for the same task to attempt to
acquire the same mutex lock once it's been locked without causing
the task to suspend itself.

The mutex that's used for this version
adds a task number entry that is recorded when the mutex lock is
first acquired. On subsequent calls to the mutex_lock() function,
if the lock is acquired, the current_task is checked against the
task that was saved and if the same, then the function simply
returns instead of stalling the task. When the sloppy_mutex lock
is released, the task number of the mutex is reset to the number
of defined tasks which ensures it can't match any task that's
attempting to acquire the lock. There is no diffennce in behavior
for different tasks which attempt to get the lock while its
held by a different task.

The motivation for this change is the nvmem module which needs to
keep a ram copy of what's in flash and the write function which uses
the ram buffer can be called multiple times prior to being done with
the ram buffer and releasing the lock.

BRANCH=none
BUG=chrome-os-partner:52520
TEST=Manual
Ran 'make runtests' then ./build/host/mutex/mutex.exe to view
console output. Verified that the tests pass and console log
matched the expected output.

Change-Id: I872242640d4c332a99f9c9c0c57930d8178992e7
Signed-off-by: Scott <scollyer@chromium.org>