Update patch set 3
Patch Set 3: Code-Review-1

This is stretching the definition of a command which is "reboot the EC" and adding a dependency on the PD module inside the EC system module.

I think we should put this in a different command.  Similar to how on the EC console, 'apreset' (reset the AP) is different than 'reboot' (reset the EC).

Then the implementation can go in common/usb_pd_protocol.c, instead of common/system.c.

Also, you're going to need some other PD-specific host commands to do the rest of the firmware update, so they might as well all go together.

Note that the ability to reboot the PD chip via this command (as well as the other commands needed for PD firmware update) should only be allowed by the EC if PD firmware update is still enabled that boot.  That is, the AP can't just ask the EC to reboot the PD at any time, and shouldn't need to ask.

Patch-set: 3