pd: Explicitly disable VCONN on CC line used for data communication
Under some circumstances, it is possible the VCONN is enabled on the CC line
used for data communication. This can happen, for example, if there was a
polarity switch and the state machine got confused. If it happens, the
affected CC line will report 5V or RP_3A mode, even if the line is
disconnected. Since VCONN is never supposed to be set on the CC line used
for the data connection, simply disable it explicitly whenever a VCONN state
change is requested.

BUG=chromium:611128,chrome-os-partner:54790
TEST=Connect and disconnect cable repeatedly. Observe CC lines when
     disconnected.

Change-Id: I28129aa9e9aa4e173fb5a5ad379d4dcb0e4e5c91
Signed-off-by: Guenter Roeck <groeck@google.com>